requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
redis==5.0.8
qdrant-client==1.11.0
sentence-transformers==3.0.1
google-generativeai==0.8.3
llama-index==0.10.45
aiohttp==3.9.5
requests==2.32.3
gunicorn==22.0.0